Overhaul to the way dungeon lockouts are stored and retrieved.
The dungeon name is no longer part of the key. GetSavedInstanceInfo has had its return parameters changed to include the dungeon name at the start. Lockout details are stored/retrieved/deleted based on the lockout ID instead of a key that combined with the name. Reset times should be more accurate.